refactor: Sections 추출 및 리팩터링
https://github.com/next-step/atdd-subway-map/pull/284#discussion_r795461347
